requirements.txt
aioredis==1.3.1
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
cloudinary==1.23.0
constantly==15.1.0
cryptography==3.1.1
daphne==3.0.1
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-cleanup==5.1.0
django-heroku==0.3.1
django-progressive-web-app==0.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.3.0
exponent-server-sdk==1.0.0
filelock==3.0.12
hiredis==1.1.0
http-ece==1.1.0
humanize==3.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
msgpack==1.0.0
msgpack-python==0.5.6
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.7
psycopg2==2.8.6
ptyprocess==0.6.0
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.9
pyfcm==1.4.7
Pygments==2.7.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.1
pyzmq==19.0.2
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==3.6.5
service-identity==18.1.0
six==1.15.0
sounddevice==0.4.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
stream-chat==1.4.0
stripe==2.53.0
suds-jurko==0.6
swapper==1.1.2.post1
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
twilio==6.47.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.28
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
xmltodict==0.12.0
zope.interface==5.1.0
